Added test Also casted the constructed max_features * n_features to an int to avoid deprecationwarnings

+++ sklearn/feature_extraction/text.py
@@ -854,7 +854,7 @@ class CountVectorizer(BaseEstimator, VectorizerMixin):
                     "max_df corresponds to < documents than min_df")
             max_features_count = (max_features
                                   if isinstance(max_features, numbers.Integral)
-                                  else max_features * n_features)
+                                  else int(max_features * n_features))
             X, self.stop_words_ = self._limit_features(X, vocabulary,
                                                        max_doc_count,
                                                        min_doc_count,
